Development

+++ mine.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 uninstall_amd_driver() {
-	echo "Uninstalling driver"
-	echo 'APT::Get::Assume-Yes "true";' >>/etc/apt/apt.conf.d/90assumeyes
-	/usr/bin/amdgpu-uninstall
-	rm /etc/apt/apt.conf.d/90assumeyes
-	echo "Done!"
+	if [ -f /usr/bin/amdgpu-uninstall ]; then
+		echo "Uninstalling driver"
+		echo 'APT::Get::Assume-Yes "true";' >>/etc/apt/apt.conf.d/90assumeyes
+		/usr/bin/amdgpu-uninstall
+		rm /etc/apt/apt.conf.d/90assumeyes
+		echo "Done!"
+	else
+		echo "AMD driver already uninstalled"
+	fi
 }
 
 install_amd_driver() {
@@ -45,6 +49,11 @@ if [[ "${INSTALLED_DRIVERV}" != "${DRIVERV}" ]]; then
 	18.20)
 		uninstall_amd_driver
 		install_amd_driver "amdgpu-pro-18.20-621984.tar.xz" "https://drivers.amd.com/drivers/linux/ubuntu-18-04" "--opencl=legacy,pal --headless"
+		;;
+
+	19.50)
+		uninstall_amd_driver
+		install_amd_driver "amdgpu-pro-19.50-967956-ubuntu-18.04.tar.xz" "https://drivers.amd.com/drivers/linux/19.50/" "--opencl=legacy,pal --headless"
 		;;
 
 	20.20)
